Step 4 Complete!
I created the undo button and had it function so when it is clicked it undoes the most recent line you drew but saves it to a special redo list.
I then created the redo button to take the most recent undone line and to put it back.
After I tested it to see it all work I then went back through and removed most of my console.logs that I had in place to see if things were working.