Step 8 Complete!
Added sticker-selection buttons and new state variables to manage the active tool
Created a drawable interface and a new sticker class to act as a command object similar to the Path class.
Refactors lines and redoStack to allow undo/redo to work seamlessly for both paths and stickers.
Updates ToolPreview to show a shadow emoji when the sticker tool is active like the pen when drawing.
Modified the mouse event listeners mousedown which now creates either a Path or Sticker based on currentTool
mousemove handles dragging for stickers.